Add files via upload
This code will prompt the user to enter a message, then use the respond function to generate a response based on the user's input. If the input matches one of the keys in the responses dictionary, the corresponding value (i.e. the bot's response) will be returned. If the input doesn't match any of the keys, the function will return a default response.

The loop will continue prompting the user for input until they enter "quit" or "exit", at which point the program will terminate.